Add max-utterance-length flag to cont_fileseg
cont_fileseg now accepts an optional parameter max-utterance-length
which cuts off audio segments at a specified amount of time (ie.
seconds).

This is most often useful when building speech recognition lattices,
as these can become corrupted when they are overly large.

A better solution would be a "soft" cut-off, where the file is split
at *roughly* max-utterance-length so as to avoid splitting in the
middle of words. This would be a best-fit solution, and would require
much more work.